ai_web_scanner/requirements.txt
streamlit==1.31.0
requests==2.31.0
beautifulsoup4==4.12.3
openai==1.10.0
pandas==2.2.0
plotly==5.18.0
reportlab==4.0.9
bcrypt==4.1.2
python-dotenv==1.0.1
lxml==5.1.0
pymongo==4.6.1
python-docx==1.1.0